Add TS nocheck results (and move to extra_experiments)
BEWARE: i'm not including a patch for eval_ts.py which removes cwd=path.parent.
i'm not including the patch because i don't know why the change was made and
don't want to break "real" experiments. But that parameter seems to break my
node:assert hack that lets me use recent typescript (which i need for //
@ts-nocheck)