Use atomic vectors in R translation whenever possible
If all elements of the list have the same type (e.g. boolean, number,
string), we should use an R vector (with the `c()` constructor instead
of `list()`).

However, if the list contains `NULL` or another list, we should use the
`list()` constructor and not `c()`.

This fix is adapted from @PootieT's suggestion. Fixes #50.