This commit reworks the execution to be friendlier to Octave.
Presumably due to a bug in Octave, it fails out with a strange
error if you attempt to run a nonworking script in a directory
different from the cwd. As such, we need to pass a cwd argument to
safe_subprocess.run in order to successfully evaluate our files.

Not that it mattered in this case, since all the scripts are
full of wrong syntax anyway.